+++ source/git/identifier.go
@@ -1,10 +1,12 @@
-package source
+package git
 
 import (
 	"net/url"
 	"path"
 	"strings"
 
+	"github.com/moby/buildkit/solver/llbsolver/provenance"
+	"github.com/moby/buildkit/source"
 	srctypes "github.com/moby/buildkit/source/types"
 	"github.com/moby/buildkit/util/sshutil"
 )
@@ -53,8 +55,40 @@ func NewGitIdentifier(remoteURL string) (*GitIdentifier, error) {
 	return &repo, nil
 }
 
-func (i *GitIdentifier) ID() string {
+func (GitIdentifier) Scheme() string {
 	return srctypes.GitScheme
+}
+
+var _ source.Identifier = (*GitIdentifier)(nil)
+
+func (id *GitIdentifier) Capture(c *provenance.Capture, pin string) error {
+	url := id.Remote
+	if id.Ref != "" {
+		url += "#" + id.Ref
+	}
+	c.AddGit(provenance.GitSource{
+		URL:    url,
+		Commit: pin,
+	})
+	if id.AuthTokenSecret != "" {
+		c.AddSecret(provenance.Secret{
+			ID:       id.AuthTokenSecret,
+			Optional: true,
+		})
+	}
+	if id.AuthHeaderSecret != "" {
+		c.AddSecret(provenance.Secret{
+			ID:       id.AuthHeaderSecret,
+			Optional: true,
+		})
+	}
+	if id.MountSSHSock != "" {
+		c.AddSSH(provenance.SSH{
+			ID:       id.MountSSHSock,
+			Optional: true,
+		})
+	}
+	return nil
 }
 
 // isGitTransport returns true if the provided str is a git transport by inspecting
