+++ source/http/source_test.go
@@ -47,7 +47,7 @@ func TestHTTPSource(t *testing.T) {
 	})
 	defer server.Close()
 
-	id := &source.HTTPIdentifier{URL: server.URL + "/foo"}
+	id := &HTTPIdentifier{URL: server.URL + "/foo"}
 
 	h, err := hs.Resolve(ctx, id, nil, nil)
 	require.NoError(t, err)
@@ -166,7 +166,7 @@ func TestHTTPDefaultName(t *testing.T) {
 	})
 	defer server.Close()
 
-	id := &source.HTTPIdentifier{URL: server.URL}
+	id := &HTTPIdentifier{URL: server.URL}
 
 	h, err := hs.Resolve(ctx, id, nil, nil)
 	require.NoError(t, err)
@@ -206,7 +206,7 @@ func TestHTTPInvalidURL(t *testing.T) {
 	server := httpserver.NewTestServer(map[string]httpserver.Response{})
 	defer server.Close()
 
-	id := &source.HTTPIdentifier{URL: server.URL + "/foo"}
+	id := &HTTPIdentifier{URL: server.URL + "/foo"}
 
 	h, err := hs.Resolve(ctx, id, nil, nil)
 	require.NoError(t, err)
@@ -236,7 +236,7 @@ func TestHTTPChecksum(t *testing.T) {
 	})
 	defer server.Close()
 
-	id := &source.HTTPIdentifier{URL: server.URL + "/foo", Checksum: digest.FromBytes([]byte("content-different"))}
+	id := &HTTPIdentifier{URL: server.URL + "/foo", Checksum: digest.FromBytes([]byte("content-different"))}
 
 	h, err := hs.Resolve(ctx, id, nil, nil)
 	require.NoError(t, err)
@@ -262,7 +262,7 @@ func TestHTTPChecksum(t *testing.T) {
 	require.Equal(t, server.Stats("/foo").AllRequests, 1)
 	require.Equal(t, server.Stats("/foo").CachedRequests, 0)
 
-	id = &source.HTTPIdentifier{URL: server.URL + "/foo", Checksum: digest.FromBytes([]byte("content-correct"))}
+	id = &HTTPIdentifier{URL: server.URL + "/foo", Checksum: digest.FromBytes([]byte("content-correct"))}
 
 	h, err = hs.Resolve(ctx, id, nil, nil)
 	require.NoError(t, err)
